requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
distlib==0.3.2
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
Pillow==8.2.0
pipenv==2021.5.29
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.7
virtualenv-clone==0.5.4
